@@ -1,34 +1,40 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
+import shortid from 'shortid';
 import Button from '../Button/Button';
 import './MessageArea.css';
 
-class MessageArea extends Component {
-  state = {
-    text: '',
+function MessageArea({ author, onSendMessage }) {
+  const [text, setText] = useState('');
+
+  const handleOnChange = event => setText(event.target.value);
+
+  const handleOnSendMessage = () => {
+    if (text.trim().length === 0) {
+      return;
+    }
+
+    onSendMessage({
+      id: shortid.generate(),
+      author,
+      text
+    });
+
+    setText('');
   };
 
-  handleOnSendMessage = () => {
-    this.props.onSendMessage({ author: this.props.author, text: this.state.text });
-    this.setState({ text: '' });
-  };
-
-  handleOnChange = event => this.setState({ text: event.target.value });
-
-  render() {
-    return (
-      <div className="message-area">
-        Send a message as <strong>{this.props.author}</strong>
-        <textarea value={this.state.text} onChange={this.handleOnChange} />
-        <Button onClick={this.handleOnSendMessage} name="Send" />
-      </div>
-    );
-  }
+  return (
+    <div className="message-area">
+      Send a message as <strong>{author}</strong>
+      <textarea value={text} onChange={handleOnChange} />
+      <Button onClick={handleOnSendMessage} name="Send" />
+    </div>
+  );
 }
 
 MessageArea.propTypes = {
   author: PropTypes.string.isRequired,
-  onSendMessage: PropTypes.func.isRequired,
+  onSendMessage: PropTypes.func.isRequired
 };
 
 export default MessageArea;
