@@ -5,7 +5,7 @@ import './AppHeader.css';
 const AppHeader = () => (
   <header className="app-header">
     <img src={footballLogo} className="app-logo" alt="logo" />
-    <h1 className="app-title">World Cup 2018 Dashboard</h1>
+    <h1 className="app-title">Women's World Cup 2019 Dashboard</h1>
   </header>
 );
 
