@@ -1,14 +1,16 @@
 import React from 'react';
-import PropTypes from 'prop-types'
+import PropTypes from 'prop-types';
 import './Button.css';
 
-const Button = ({onClick, name}) => (
-  <button className="button" onClick={onClick}>{name}</button>
+const Button = ({ onClick, name, ...props }) => (
+  <button className="button" onClick={onClick} {...props}>
+    {name}
+  </button>
 );
 
 Button.propTypes = {
   name: PropTypes.string.isRequired,
-  onClick: PropTypes.func.isRequired,
+  onClick: PropTypes.func.isRequired
 };
 
 export default Button;
